tests fixed

+++ src/main/java/main/lab1/exceptions/TaskAlreadyExistsException.java
@@ -1,7 +1,7 @@
 package main.lab1.exceptions;
 
 public class TaskAlreadyExistsException extends RuntimeException {
-  public TaskAlreadyExistsException(int id) {
+  public TaskAlreadyExistsException(long id) {
     super("Task with id = " + id + " already exists.");
   }
 }
