+++ src/main/java/main/lab1/services/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 @Service
 public interface UserService {
 
-    User getUserById(int id);
+    User getUserById(long id);
 
     void createUser(User user);
